cts/app/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-bootstrap-form-horizontal==0.0.4
django-ckeditor==6.5.1
django-debug-toolbar==4.0.0
django-embed-video==1.4.8
django-haystack==3.2.1
django-js-asset==2.0.0
django-phone-field==1.8.1
django-phonenumber-field==7.0.2
django-postgres-extra==2.0.8
django-richtextfield==1.6.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
phonenumbers==8.13.9
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
urllib3==1.26.15
